refactor: extract volume polling into services/volume.ts
Polling logic (pollVolume, pollMute, interval, VolumeError) extracted
from main/index.ts into main/services/volume.ts. The function receives
AppState and a callbacks object with onFatalError, decoupling it from
direct app.exit() calls.

Also fixes TaggedError class definition (missing empty call before
class body) caught by tsgo's stricter type checking.
